Move config files to optuna_dashboard_client
