+++ optuna_dashboard_client/package.json
@@ -1,0 +1,67 @@
+{
+  "name": "optuna-dashboard",
+  "private": true,
+  "version": "0.0.1",
+  "description": "Dashboard for Optuna",
+  "main": "index.js",
+  "scripts": {
+    "watch": "NODE_ENV=development TYPESCRIPT_LOADER=esbuild-loader webpack --watch",
+    "build": "webpack",
+    "build:dev": "NODE_ENV=development TYPESCRIPT_LOADER=esbuild-loader webpack",
+    "build:prd": "NODE_ENV=production webpack",
+    "test": "jest tests"
+  },
+  "author": "Masashi Shibata",
+  "license": "MIT",
+  "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.6",
+    "@mui/lab": "^5.0.0-alpha.162",
+    "@mui/material": "^5.15.6",
+    "@react-three/drei": "^9.96.4",
+    "@react-three/fiber": "^8.15.15",
+    "@tanstack/react-virtual": "^3.1.2",
+    "@tanstack/react-query": "^5.18.1",
+    "@types/three": "^0.160.0",
+    "axios": "^1.6.7",
+    "elkjs": "^0.9.1",
+    "notistack": "^3.0.1",
+    "plotly.js-dist-min": "^2.28.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^9.0.1",
+    "react-router-dom": "^6.21.3",
+    "react-syntax-highlighter": "^15.5.0",
+    "reactflow": "^11.10.3",
+    "recoil": "^0.7.7",
+    "rehype-mathjax": "^6.0.0",
+    "rehype-raw": "^7.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-math": "^6.0.0",
+    "three": "^0.160.1",
+    "wavesurfer.js": "^7.7.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.23.9",
+    "@babel/preset-env": "^7.23.9",
+    "@testing-library/react": "^14.1.2",
+    "@types/jest": "^29.5.11",
+    "@types/plotly.js": "^2.12.32",
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
+    "@types/react-syntax-highlighter": "^15.5.11",
+    "compression-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.9.1",
+    "esbuild-loader": "^4.0.3",
+    "jest": "^29.7.0",
+    "jest-canvas-mock": "^2.5.2",
+    "jest-environment-jsdom": "^29.7.0",
+    "style-loader": "^3.3.4",
+    "ts-jest": "^29.1.2",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.3.3",
+    "webpack": "^5.90.0",
+    "webpack-cli": "^5.1.4"
+  }
+}
